Moved `downloadToFile` to utils
